% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoAuth.R
\name{AmoAuth}
\alias{AmoAuth}
\title{Authorization}
\usage{
AmoAuth(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  verbose = T
)
}
\arguments{
\item{email}{Your email in AmoCRM, check xxx.amocrm.ru/settings/profile/}

\item{apikey}{Your api key from settings in interface, check xxx.amocrm.ru/settings/profile/}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{verbose}{Printing the answer}
}
\value{
TRUE if ok and error if not
}
\description{
Auth, using in other functions
}
\examples{
AmoAuth(email = 'test@test.ru', apikey = 'jadbfjeabfkj', domain = 'test')
}
