% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoAuthList.R
\name{AmoAuthList}
\alias{AmoAuthList}
\title{Auth list}
\usage{
AmoAuthList(email = NULL, apikey = NULL, domain = NULL)
}
\arguments{
\item{email}{Your email in AmoCRM, check xxx.amocrm.ru/settings/profile/}

\item{apikey}{Your api key from settings in interface, check xxx.amocrm.ru/settings/profile/}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}
}
\value{
List of auth parameters
}
\description{
Using for nice-passing auth parameters
}
\examples{
\dontrun{
AmoAuthList("email@email.ru", "shdfuisgea3624dwe6t", "test")
}
}
