% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoCustomers.R
\name{AmoCustomers}
\alias{AmoCustomers}
\title{Customers}
\usage{
AmoCustomers(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  limit = 500,
  flatten = F,
  id = NULL,
  query = NULL,
  main_user = NULL,
  date_type = NULL,
  date_from = NULL,
  date_to = NULL,
  next_date_from = NULL,
  next_date_to = NULL
)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{limit}{Batch limit, sometimes AmoCRM's API doesn't work properly, you can reduce the value and have a chance to load your data}

\item{flatten}{Set \code{TRUE} if you want to join all the output dataframes You'll have a not tidy-dataframe with left-joining all dataframes}

\item{id}{Filter. Pass id or vector of ids of customers.}

\item{query}{Filter. Searching for all fields of customers. String.}

\item{main_user}{Filter. Pass id or vector of ids of responsible user ids. You can get ids from AmoUsers().}

\item{date_type}{Filter. Choose date type which you want to filter (\code{'create'} or \code{'modify'})}

\item{date_from}{Filter. Date from, e.g. \code{'2019-01-01'}.}

\item{date_to}{Filter. Date to, e.g. \code{'2019-01-01'}.}

\item{next_date_from}{Filter. Date from of next purchasing.}

\item{next_date_to}{Filter. Date to of next purchasing.}
}
\value{
If flatten is \code{FALSE} (default) you'll get a list of 4 tidy-dataframes which you can join by id. You can access it using list_name$dataframe_name.

linked_custom_fields — linked custom fields with all parameters.

linked_tags — linked tags with all parameters.

linked_contacts — linked contacts with all parameters.
}
\description{
Function to get customers.
}
\examples{
library(dplyr)
customers <- AmoCustomers(auth_list = auth_list)
customers_with_cf <- customers$customers \%>\%
                        left_join(customers$linked_custom_fields, by = 'id') # not tidy
}
\references{
Please \strong{READ} this:
\href{https://github.com/grkhr/amocrm/blob/master/md/AmoCustomers.md}{Function documentation in Russian on GitHub}

Also nice to read:
\href{https://www.amocrm.ru/developers/content/api/customers}{AmoCRM official documentation}
}
