% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoCustomersTransactions.R
\name{AmoCustomersTransactions}
\alias{AmoCustomersTransactions}
\title{Customers transactions}
\usage{
AmoCustomersTransactions(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  limit = 500,
  id = NULL,
  customer_id = NULL
)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{limit}{Batch limit, sometimes AmoCRM's API doesn't work properly, you can reduce the value and have a chance to load your data}

\item{id}{Filter. Pass id of transaction.}

\item{customer_id}{Filter. Pass id of customer to get transaction from its. You can get ids from AmoCustomers().}
}
\value{
Dataframe in output.
}
\description{
Function to get transactions of customers.
}
\examples{
\dontrun{
# get all
transactions <- AmoCustomersTransations(auth_list = auth_list)
}

}
\references{
Please \strong{READ} this:
\href{https://github.com/grkhr/amocrm/blob/master/md/AmoCustomersTransactions.md}{Function documentation in Russian on GitHub}

Also nice to read:
\href{https://www.amocrm.ru/developers/content/api/customers}{AmoCRM official documentation}
}
