% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoCustomFields.R
\name{AmoCustomFields}
\alias{AmoCustomFields}
\title{Custom fields}
\usage{
AmoCustomFields(email = NULL, apikey = NULL, domain = NULL, auth_list = NULL)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}
}
\value{
You'll get a list of list of dataframes which you can join by id.

list$custom_fields_xxx$custom_fields - ids and names

list$custom_fields_xxx$custom_fields_enum - multiple values
}
\description{
Function to get custom field names and ids for all. Please read the following manual on github: \code{\link{https://github.com/grkhr/amocrm}}
}
\details{
Check api params if needed: \code{\link{https://www.amocrm.ru/developers/content/api/account}}
}
\examples{
library(dplyr)
custom_fields <- AmoCustomFields(auth_list = auth_list)
custom_fields_contacts_with_enum <- custom_field$custom_fields_contacts$custom_fields \%>\%
  left_join(custom_field$custom_fields_contacts$custom_fields_enum, by = 'id') # not tidy
}
