% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoGroups.R
\name{AmoGroups}
\alias{AmoGroups}
\title{Groups}
\usage{
AmoGroups(email = NULL, apikey = NULL, domain = NULL, auth_list = NULL)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}
}
\value{
You'll get a groups.
}
\description{
Function to get groups of users. Please read the following manual on github: \code{\link{https://github.com/grkhr/amocrm}}
}
\examples{
groups <- AmoGroups(auth_list = auth_list)
}
