% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoIncomingLeads.R
\name{AmoIncomingLeads}
\alias{AmoIncomingLeads}
\title{Incoming Leads}
\usage{
AmoIncomingLeads(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  limit = 500,
  categories = NULL,
  order_by_key = NULL,
  order_by_value = NULL,
  pipeline_id = NULL
)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{limit}{Batch limit, sometimes AmoCRM's API doesn't work properly, you can reduce the value and have a chance to load your data}

\item{categories}{Filter. Categories. Vector or single value of \code{"sip"}, \code{"mail"}, \code{"forms"} or \code{"chats"}.}

\item{order_by_key}{Filter. Key of ordering. For example \code{"created_at"}.}

\item{order_by_value}{Filter. \code{"asc"} or \code{"desc"}. Works if \code{order_by_key} is set.}

\item{pipeline_id}{Filter. You can pass single id and get imcoming leads only from this pipeline. You can get ids from AmoPipelinesStatuses().}
}
\value{
Dataframe in output.
}
\description{
Function to get incoming leads.
}
\examples{
# get all
incleads <- AmoIncomingLeads(auth_list = auth_list)

# filtered
incleads <- AmoIncomingLeads(auth_list = auth_list,
                             categories = c('sip','mail'),
                             order_by_key = 'created_at',
                             order_by_value = 'desc')

}
\references{
Please \strong{READ} this:
\href{https://github.com/grkhr/amocrm/blob/master/md/AmoIncomingLeads.md}{Function documentation in Russian on GitHub}

Also nice to read:
\href{https://www.amocrm.ru/developers/content/api/unsorted}{AmoCRM official documentation}
}
