% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoLeads.R
\name{AmoLeads}
\alias{AmoLeads}
\title{Leads}
\usage{
AmoLeads(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  limit = 500,
  flatten = F,
  id = NULL,
  query = NULL,
  responsible_user_id = NULL,
  with_with = "is_price_modified_by_robot,loss_reason_name",
  status = NULL,
  date_create_from = NULL,
  date_create_to = NULL,
  date_modify_from = NULL,
  date_modify_to = NULL,
  tasks = NULL,
  active = NULL,
  timezone = 3
)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{limit}{Batch limit, sometimes AmoCRM's API doesn't work properly, you can reduce the value and have a chance to load your data}

\item{flatten}{Set TRUE if you want to join all the output dataframes You'll have a not tidy-dataframe with left-joining all dataframes}

\item{id}{Filter. Pass id or vector of ids of leads.}

\item{query}{Filter. Searching for all fields of leads. String.}

\item{responsible_user_id}{Filter. Pass id or vector of ids of responsible user ids. You can get ids from AmoUsers().}

\item{with_with}{Additional data. Default to 'is_price_modified_by_robot,loss_reason_name'.}

\item{status}{Filter. Single status id or vector of ids. You can get ids from AmoPipelinesStatuses().}

\item{date_create_from}{Filter. Date create of lead. You can pass like '2019-01-01' or with time like '2019-01-01 12:00:00'}

\item{date_create_to}{Filter. Date create of lead. You can pass like '2019-01-01' or with time like '2019-01-01 12:00:00'}

\item{date_modify_from}{Filter. Date modify of lead. You can pass like '2019-01-01' or with time like '2019-01-01 12:00:00'}

\item{date_modify_to}{Filter. Date modify of lead. You can pass like '2019-01-01' or with time like '2019-01-01 12:00:00'}

\item{tasks}{Filter. Pass 1 if you need leads without tasks, pass 2 if you need leads with undone tasks.}

\item{active}{Filter. Pass 1 if you need only active leads.}
}
\value{
If flatten is F (default) you'll get a list of 4 tidy-dataframes which you can join by id. You can access it using list_name$dataframe_name.

leads - all leads with unnested parameters.

linked_custom_fields — linked custom fields with all parameters.

linked_tags — linked tags with all parameters.

linked_contacts — linked contacts with all parameters.
}
\description{
Function to get leads. Please read the following manual on github: \code{\link{https://github.com/grkhr/amocrm}}
}
\details{
Check api params if needed: \code{\link{https://www.amocrm.ru/developers/content/api/leads}}
}
\examples{
# simple
library(dplyr)
leads <- AmoLeads(auth_list = auth_list)
leads_with_cf <- leads$leads \%>\%
                        left_join(leads$linked_custom_fields, by = 'id') # not tidy

# filters
leads <- AmoLeads(auth_list = auth_list,
                  date_create_from = '2019-02-01 05:00:00',
                  date_create_to = '2019-02-20 17:00:00',
                  active = 1)
}
