% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoNotes.R
\name{AmoNotes}
\alias{AmoNotes}
\title{Notes}
\usage{
AmoNotes(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  limit = 500,
  id = NULL,
  element_id = NULL,
  type = "contact",
  note_type = NULL,
  if_modified_since = NULL,
  all = F
)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{limit}{Batch limit, sometimes AmoCRM's API doesn't work properly, you can reduce the value and have a chance to load your data}

\item{id}{Filter. Pass id or vector of ids of notes.}

\item{element_id}{Filter. Id of lead/contact/etc. Vector is not allowed.}

\item{type}{What to get. You can pass \code{"contact"}, \code{"lead"}, \code{"company"} or \code{"task"}. Default to \code{"contact"}. If you need all, look at \code{"all"} parameter.}

\item{note_type}{Type of note. Useful filter. Check \href{https://github.com/grkhr/amocrm/blob/master/md/AmoNotes.md#note_type}{all note types}.}

\item{if_modified_since}{Filter. Get notes after some timestamp. Pass time like \code{'2019-01-01 12:30:00'}.}

\item{all}{If you want to load all note for all types, set TRUE. You'll get list of dataframes.}
}
\value{
Dataframe in output (or list of dataframes if \code{all = TRUE}.)
}
\description{
Function to get notes.
}
\examples{
\dontrun{
# leads
notes <- AmoNotes(aiuth_list = auth_list, type = 'lead')

# get lead's statuses changes from 1 june 2019
notes <- AmoNotes(auth_list = auth_list, type = 'lead', note_type = 3, if_modified_since = '2019-06-01 00:00:00')
}
}
\references{
Please \strong{READ} this:
\href{https://github.com/grkhr/amocrm/blob/master/md/AmoNotes.md}{Function documentation in Russian on GitHub}

Also nice to read:
\href{https://www.amocrm.ru/developers/content/api/notes}{AmoCRM official documentation}
}
