% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoTasks.R
\name{AmoTasks}
\alias{AmoTasks}
\title{Tasks}
\usage{
AmoTasks(
  email = NULL,
  apikey = NULL,
  domain = NULL,
  auth_list = NULL,
  limit = 500,
  id = NULL,
  type = NULL,
  element_id = NULL,
  responsible_user_id = NULL,
  date_create_from = NULL,
  date_create_to = NULL,
  date_modify_from = NULL,
  date_modify_to = NULL,
  status = NULL,
  created_by = NULL,
  task_type = NULL
)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}

\item{limit}{Batch limit, sometimes AmoCRM's API doesn't work properly, you can reduce the value and have a chance to load your data}

\item{id}{Filter. Pass id or vector of ids of tasks.}

\item{type}{Filter. Works if element_id is set. Pass "lead", "contact", "company" or "customer".}

\item{element_id}{Filter. Pass contact/lead/etc id.}

\item{responsible_user_id}{Filter. Pass id or vector of ids of responsible user ids. You can get ids from AmoUsers().}

\item{date_create_from}{Filter. Date create of taks. You can pass like '2019-01-01' or with time in UTC timezone like '2019-01-01 12:00:00'}

\item{date_create_to}{Filter. Date create of taks. You can pass like '2019-01-01' or with time in UTC timezone like '2019-01-01 12:00:00'}

\item{date_modify_from}{Filter. Date modify of taks. You can pass like '2019-01-01' or with time in UTC timezone like '2019-01-01 12:00:00'}

\item{date_modify_to}{Filter. Date modify of taks. You can pass like '2019-01-01' or with time in UTC timezone like '2019-01-01 12:00:00'}

\item{status}{Filter. Pass 1 if you need done tasks, pass 0 if undone tasks.}

\item{created_by}{Filter. Tasks by author. Pass if of user or vector of ids.}

\item{task_type}{Filter. Task by its type. Pass id. You can get id from AmoTaskTypes().}
}
\value{
Dataframe in output.
}
\description{
Function to get tasks. Please read the following manual on github: \code{\link{https://github.com/grkhr/amocrm}}
}
\details{
Check api params if needed: \code{\link{https://www.amocrm.ru/developers/content/api/tasks}}
}
\examples{
# simple
tasks <- AmoTasks(auth_list = auth_list)

# filters
tasks <- AmoTasks(auth_list = auth_list,
                  type = 'lead',
                  date_create_from = '2019-02-01 05:00:00',
                  date_create_to = '2019-02-20 17:00:00',
                  status = 0,
                  task_type = 1)
}
