% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoTaskTypes.R
\name{AmoTaskTypes}
\alias{AmoTaskTypes}
\title{Task types from account}
\usage{
AmoTaskTypes(email = NULL, apikey = NULL, domain = NULL, auth_list = NULL)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}
}
\value{
Dataframe in output.
}
\description{
Function to get tasks.
}
\examples{
\dontrun{
task_types <- AmoTaskTypes(auth_list = auth_list)
}
}
\references{
Please \strong{READ} this:
\href{https://github.com/grkhr/amocrm/blob/master/md/AmoTaskTypes.md}{Function documentation in Russian on GitHub}

Also nice to read:
\href{https://www.amocrm.ru/developers/content/api/tasks#type}{AmoCRM official documentation}
}
