% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmoTaskTypes.R
\name{AmoTaskTypes}
\alias{AmoTaskTypes}
\title{Task types from account}
\usage{
AmoTaskTypes(email = NULL, apikey = NULL, domain = NULL, auth_list = NULL)
}
\arguments{
\item{email}{Email}

\item{apikey}{Your api key from settings in interface}

\item{domain}{Your domain in AmoCRM (xxx in xxx.amocrm.ru)}

\item{auth_list}{List with auth data, you can build from AmoAuthList}
}
\value{
Dataframe in output.
}
\description{
Function to get tasks. Please read the following manual on github: \code{\link{https://github.com/grkhr/amocrm}}
}
\details{
Check api params if needed: \code{\link{https://www.amocrm.ru/developers/content/api/account}}
}
\examples{
task_types <- AmoTaskTypes(auth_list = auth_list)
}
